libs/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genblaze-core"
dynamic = ["version"]
description = "Core SDK for genblaze media generation orchestration"
readme = "../../README.md"
requires-python = ">=3.11"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Multimedia",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.0,<3",
    "pillow>=10.0",
]

[project.urls]
Documentation = "https://genblaze.dev"
Repository = "https://github.com/backblaze-b2-samples/genblaze"
Changelog = "https://github.com/backblaze-b2-samples/genblaze/blob/main/CHANGELOG.md"
Issues = "https://github.com/backblaze-b2-samples/genblaze/issues"

[project.optional-dependencies]
parquet = ["pyarrow>=14.0"]
audio = ["mutagen>=1.47"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "pyarrow>=14.0",
    "mutagen>=1.47",
    "jsonschema>=4.0",
    "hypothesis>=6.0",
]

[tool.hatch.version]
path = "genblaze_core/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["genblaze_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.coverage.run]
source = ["genblaze_core"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 70
